+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ This document describes changes between each past release.
   **Important**: Because of a limitation in `Pyramid tooling <http://stackoverflow.com/a/21228232/147077>`_,
   it won't work if the port is hard-coded in your existing ``.ini`` file. Replace
   it by ``%(http_port)s`` or regenerate a new configuration file with ``kinto init``.
+- Add support for ``pool_timeout`` option in Redis backend (fixes #620)
 
 
 3.1.0 (2016-05-24)